requirements.txt
python-dotenv
pandas
numpy
matplotlib
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
mlflow
pyyaml


##-e .